add options group and categorys. alterated routes

+++ src/3.controller/class/User/3.AdmUser.js
@@ -2,12 +2,13 @@ import GuestUser from "./2.GuestUser.js";
 import errorHandling from "../../../2.service/errorHandling/errorHandling.js";
 import chalk from "chalk";
 import userService from "../../../2.service/busnessRoule/user/userService.js";
-import DataGroupValitadion from "../../valitadtion/group/DataGroupValitadion.js";
+import groupAndCategoryValdiation from "../../valitadtion/groupAndCategory/groupAndCategoryValdiation.js";
 import Event from "../Event/Event.js";
 
 class AdmUser extends GuestUser {
   createEvent = async (reqBody) => {
-    const event = new Event(reqBody) 
+    const event = new Event(reqBody);
+    return event.add(reqBody)
   };
 
   // editEvent = async (reqBody) => {};
@@ -18,13 +19,13 @@ class AdmUser extends GuestUser {
 
   createGroup = async (reqBody) => {
     try {
-      const dataValidation = DataGroupValitadion(reqBody);
-
+      console.log("Passando por: createGroup")
+      const dataValidation = await groupAndCategoryValdiation(reqBody);
+      console.log("O status é:" +dataValidation.status);
       if (dataValidation.status) {
         return await userService.createGroup(reqBody);
-      } else {
-        throw dataValidation;
       }
+      throw dataValidation;
     } catch (error) {
       return errorHandling(error);
     }
@@ -32,50 +33,64 @@ class AdmUser extends GuestUser {
 
   editGroup = async (reqBody) => {
     try {
-      return await userService.editGroup(reqBody);
+      const dataValidation = groupAndCategoryValdiation(reqBody);
+      if (dataValidation.status) {
+        return await userService.editGroup(reqBody);
+      }
+      throw dataValidation;
     } catch (error) {
       return errorHandling(error);
     }
   };
 
-  deleteGroup = async () => {
+  deleteGroup = async (reqBody) => {
     try {
-      return await userService.deleteGroup();
+      const dataValidation = groupAndCategoryValdiation(reqBody);
+      if (dataValidation.status) {
+        return await userService.deleteGroup();
+      }
+      throw dataValidation;
     } catch (error) {
       return errorHandling(error);
     }
   };
 
-  // CATEGORIA (NADA EDITADO ABAIXO)
-  // createCategory = async () => {
-  //   try {
+  createCategory = async (reqBody) => {
+    try {
+      const dataValidation = groupAndCategoryValdiation(reqBody);
 
-  //     return await userService.deleteGroup();
+      if (dataValidation.status) {
+        return await userService.createCategory(reqBody);
+      }
+      throw dataValidation;
+    } catch (error) {
+      return errorHandling(error);
+    }
+  };
 
-  // } catch (error) {
-  //   return errorHandling(error);
-  // }
-  // };
+  editCategory = async (reqBody) => {
+    try {
+      const dataValidation = groupAndCategoryValdiation(reqBody);
+      if (dataValidation.status) {
+        return await userService.editCategory(reqBody);
+      }
+      throw dataValidation;
+    } catch (error) {
+      return errorHandling(error);
+    }
+  };
 
-  // editCategory = async (reqBody) => {
-  //   try {
-
-  //       return await userService.editGroup(reqBody);
-
-  //   } catch (error) {
-  //     return errorHandling(error);
-  //   }
-  // };
-
-  // deleteCategory = async (reqBody) => {
-  //   try {
-
-  //       return await userService.editGroup(reqBody);
-
-  //   } catch (error) {
-  //     return errorHandling(error);
-  //   }
-  // };
+  deleteCategory = async (reqBody) => {
+    try {
+      const dataValidation = groupAndCategoryValdiation(reqBody);
+      if (dataValidation.status) {
+        return await userService.deleteCategory();
+      }
+      throw dataValidation;
+    } catch (error) {
+      return errorHandling(error);
+    }
+  };
 }
 
 export default AdmUser;
